requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
apscheduler==3.10.4
python-telegram-bot==21.9
bcrypt==4.2.1
pyyaml==6.0.2
aiofiles==24.1.0
itsdangerous==2.2.0
python-multipart==0.0.20
jinja2==3.1.5
pytz==2024.2
requests==2.32.3
python-dateutil==2.9.0
